@@ -22,16 +22,16 @@
     "release": "nx release --skip-publish"
   },
   "dependencies": {
-    "@angular/animations": "19.1.4",
-    "@angular/common": "19.1.4",
-    "@angular/compiler": "19.1.4",
-    "@angular/core": "19.1.4",
-    "@angular/forms": "19.1.4",
-    "@angular/platform-browser": "19.1.4",
-    "@angular/platform-browser-dynamic": "19.1.4",
-    "@angular/router": "19.1.4",
-    "@nx/angular": "20.4.0",
-    "@nx/devkit": "20.4.0",
+    "@angular/animations": "20.0.4",
+    "@angular/common": "20.0.4",
+    "@angular/compiler": "20.0.4",
+    "@angular/core": "20.0.4",
+    "@angular/forms": "20.0.4",
+    "@angular/platform-browser": "20.0.4",
+    "@angular/platform-browser-dynamic": "20.0.4",
+    "@angular/router": "20.0.4",
+    "@nx/angular": "21.2.1",
+    "@nx/devkit": "21.2.1",
     "markdown-table": "3.0.4",
     "rxjs": "7.8.1",
     "ts-dedent": "2.2.0",
@@ -39,49 +39,49 @@
     "zone.js": "0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "19.1.5",
-    "@angular-devkit/core": "19.1.5",
-    "@angular-devkit/schematics": "19.1.5",
-    "@angular-eslint/eslint-plugin": "19.0.2",
-    "@angular-eslint/eslint-plugin-template": "19.0.2",
-    "@angular-eslint/template-parser": "19.0.2",
-    "@angular/cli": "19.1.5",
-    "@angular/compiler-cli": "19.1.4",
-    "@angular/language-service": "19.1.4",
+    "@angular-devkit/build-angular": "20.0.3",
+    "@angular-devkit/core": "20.0.3",
+    "@angular-devkit/schematics": "20.0.3",
+    "@angular-eslint/eslint-plugin": "20.0.0",
+    "@angular-eslint/eslint-plugin-template": "20.0.0",
+    "@angular-eslint/template-parser": "20.0.0",
+    "@angular/cli": "~20.0.0",
+    "@angular/compiler-cli": "20.0.4",
+    "@angular/language-service": "20.0.4",
     "@compodoc/compodoc": "1.1.26",
-    "@nx/cypress": "20.4.0",
-    "@nx/eslint": "20.4.0",
-    "@nx/eslint-plugin": "20.4.0",
-    "@nx/jest": "20.4.0",
-    "@nx/js": "20.4.0",
-    "@nx/node": "20.4.0",
-    "@nx/web": "20.4.0",
-    "@nx/workspace": "20.4.0",
-    "@schematics/angular": "19.1.5",
+    "@nx/cypress": "21.2.1",
+    "@nx/eslint": "21.2.1",
+    "@nx/eslint-plugin": "21.2.1",
+    "@nx/jest": "21.2.1",
+    "@nx/js": "21.2.1",
+    "@nx/node": "21.2.1",
+    "@nx/web": "21.2.1",
+    "@nx/workspace": "21.2.1",
+    "@schematics/angular": "20.0.3",
     "@swc-node/register": "1.10.9",
     "@swc/core": "1.10.12",
     "@swc/helpers": "0.5.15",
     "@types/jest": "29.5.14",
     "@types/node": "22.13.0",
-    "@typescript-eslint/eslint-plugin": "8.22.0",
-    "@typescript-eslint/parser": "8.22.0",
-    "@typescript-eslint/utils": "8.22.0",
+    "@typescript-eslint/eslint-plugin": "8.29.0",
+    "@typescript-eslint/parser": "8.29.0",
+    "@typescript-eslint/utils": "8.29.0",
     "@yandex/ymaps3-types": "1.0.15731128",
     "autoprefixer": "10.4.20",
     "concurrently": "9.1.2",
     "docsify-cli": "4.4.4",
     "eslint": "8.57.0",
-    "eslint-config-prettier": "9.1.0",
+    "eslint-config-prettier": "10.0.0",
     "eslint-plugin-cypress": "3.6.0",
     "eslint-plugin-simple-import-sort": "12.1.1",
     "husky": "9.1.7",
     "jest": "29.7.0",
     "jest-environment-jsdom": "29.7.0",
     "jest-environment-node": "29.7.0",
-    "jest-preset-angular": "14.5.1",
+    "jest-preset-angular": "14.6.0",
     "jsonc-eslint-parser": "2.4.0",
-    "ng-packagr": "19.1.2",
-    "nx": "20.4.0",
+    "ng-packagr": "20.0.1",
+    "nx": "21.2.1",
     "postcss": "8.5.1",
     "postcss-import": "16.1.0",
     "postcss-preset-env": "10.1.3",
@@ -89,7 +89,7 @@
     "prettier": "3.4.2",
     "ts-jest": "29.2.5",
     "ts-node": "10.9.2",
-    "typescript": "5.7.3",
+    "typescript": "5.8.3",
     "wait-on": "8.0.2"
   }
 }
