@@ -1,6 +1,7 @@
 import {
   Directive,
   EventEmitter,
+  inject,
   Input,
   NgZone,
   OnChanges,
@@ -59,6 +60,9 @@ import { YMapComponent } from '../y-map/y-map.component';
   standalone: true,
 })
 export class YMapFeatureDirective implements OnInit, OnDestroy, OnChanges {
+  private readonly ngZone = inject(NgZone);
+  private readonly yMapComponent = inject(YMapComponent);
+
   private readonly destroy$ = new Subject<void>();
 
   private feature?: YMapFeature;
@@ -75,11 +79,6 @@ export class YMapFeatureDirective implements OnInit, OnDestroy, OnChanges {
   @Output() ready: EventEmitter<YReadyEvent<YMapFeature>> = new EventEmitter<
     YReadyEvent<YMapFeature>
   >();
-
-  constructor(
-    private readonly ngZone: NgZone,
-    private readonly yMapComponent: YMapComponent,
-  ) {}
 
   ngOnInit() {
     this.yMapComponent.map$.pipe(filter(Boolean), takeUntil(this.destroy$)).subscribe((map) => {
