@@ -1,6 +1,7 @@
 import {
   Directive,
   EventEmitter,
+  inject,
   Input,
   NgZone,
   OnChanges,
@@ -44,6 +45,9 @@ import { YMapControlsDirective } from '../y-map-controls/y-map-controls.directiv
   standalone: true,
 })
 export class YMapControlButtonDirective implements OnInit, OnChanges, OnDestroy {
+  private readonly ngZone = inject(NgZone);
+  private readonly yMapControlsDirective = inject(YMapControlsDirective);
+
   private readonly destroy$ = new Subject<void>();
 
   private control?: YMapControlButton;
@@ -65,11 +69,6 @@ export class YMapControlButtonDirective implements OnInit, OnChanges, OnDestroy 
   @Output() ready: EventEmitter<YReadyEvent<YMapControlButton>> = new EventEmitter<
     YReadyEvent<YMapControlButton>
   >();
-
-  constructor(
-    private readonly ngZone: NgZone,
-    private readonly yMapControlsDirective: YMapControlsDirective,
-  ) {}
 
   ngOnInit() {
     this.yMapControlsDirective.controls$
