@@ -1,6 +1,7 @@
 import {
   Directive,
   EventEmitter,
+  inject,
   Input,
   NgZone,
   OnChanges,
@@ -42,6 +43,9 @@ import { YMapComponent } from '../../common/y-map/y-map.component';
   standalone: true,
 })
 export class YMapControlsDirective implements OnInit, OnChanges, OnDestroy {
+  private readonly ngZone = inject(NgZone);
+  private readonly yMapComponent = inject(YMapComponent);
+
   private readonly destroy$ = new Subject<void>();
 
   controls$ = new BehaviorSubject<YMapControls | null>(null);
@@ -63,11 +67,6 @@ export class YMapControlsDirective implements OnInit, OnChanges, OnDestroy {
   @Output() ready: EventEmitter<YReadyEvent<YMapControls>> = new EventEmitter<
     YReadyEvent<YMapControls>
   >();
-
-  constructor(
-    private readonly ngZone: NgZone,
-    private readonly yMapComponent: YMapComponent,
-  ) {}
 
   ngOnInit() {
     this.yMapComponent.map$.pipe(filter(Boolean), takeUntil(this.destroy$)).subscribe((map) => {
