@@ -1,6 +1,7 @@
 import {
   Directive,
   EventEmitter,
+  inject,
   Input,
   NgZone,
   OnChanges,
@@ -46,6 +47,9 @@ import { YMapControlsDirective } from '../y-map-controls/y-map-controls.directiv
   standalone: true,
 })
 export class YMapOpenMapsButtonDirective implements OnInit, OnChanges, OnDestroy {
+  private readonly ngZone = inject(NgZone);
+  private readonly yMapControlsDirective = inject(YMapControlsDirective);
+
   private readonly destroy$ = new Subject<void>();
 
   private control?: YMapOpenMapsButton;
@@ -62,11 +66,6 @@ export class YMapOpenMapsButtonDirective implements OnInit, OnChanges, OnDestroy
   @Output() ready: EventEmitter<YReadyEvent<YMapOpenMapsButton>> = new EventEmitter<
     YReadyEvent<YMapOpenMapsButton>
   >();
-
-  constructor(
-    private readonly ngZone: NgZone,
-    private readonly yMapControlsDirective: YMapControlsDirective,
-  ) {}
 
   ngOnInit() {
     this.yMapControlsDirective.controls$
