@@ -1,6 +1,7 @@
 import {
   Directive,
   EventEmitter,
+  inject,
   Input,
   NgZone,
   OnChanges,
@@ -39,6 +40,9 @@ import { YMapComponent } from '../../common/y-map/y-map.component';
   standalone: true,
 })
 export class YMapDefaultSchemeLayerDirective implements OnInit, OnDestroy, OnChanges {
+  private readonly ngZone = inject(NgZone);
+  private readonly yMapComponent = inject(YMapComponent);
+
   private readonly destroy$ = new Subject<void>();
 
   private layer?: YMapDefaultSchemeLayer;
@@ -60,11 +64,6 @@ export class YMapDefaultSchemeLayerDirective implements OnInit, OnDestroy, OnCha
   @Output() ready: EventEmitter<YReadyEvent<YMapDefaultSchemeLayer>> = new EventEmitter<
     YReadyEvent<YMapDefaultSchemeLayer>
   >();
-
-  constructor(
-    private readonly ngZone: NgZone,
-    private readonly yMapComponent: YMapComponent,
-  ) {}
 
   ngOnInit() {
     this.yMapComponent.map$.pipe(filter(Boolean), takeUntil(this.destroy$)).subscribe((map) => {
